Refactor of the tasks
