+++ cognee/api/v1/cognify/cognify_v2.py
@@ -9,14 +9,6 @@ from cognee.modules.data.processing.document_types.AudioDocument import AudioDoc
 from cognee.modules.data.processing.document_types.ImageDocument import ImageDocument
 from cognee.shared.data_models import KnowledgeGraph
 from cognee.modules.data.processing.document_types import PdfDocument, TextDocument
-# from cognee.modules.cognify.vector import save_data_chunks
-# from cognee.modules.data.processing.process_documents import process_documents
-# from cognee.modules.classification.classify_text_chunks import classify_text_chunks
-# from cognee.modules.data.extraction.data_summary.summarize_text_chunks import summarize_text_chunks
-# from cognee.modules.data.processing.filter_affected_chunks import filter_affected_chunks
-# from cognee.modules.data.processing.remove_obsolete_chunks import remove_obsolete_chunks
-# from cognee.modules.data.extraction.knowledge_graph.expand_knowledge_graph import expand_knowledge_graph
-# from cognee.modules.data.extraction.knowledge_graph.establish_graph_topology import establish_graph_topology
 from cognee.modules.data.models import Dataset, Data
 from cognee.modules.data.operations.get_dataset_data import get_dataset_data
 from cognee.modules.data.operations.retrieve_datasets import retrieve_datasets
@@ -31,6 +23,7 @@ from cognee.tasks.chunk_extract_summary.chunk_extract_summary import chunk_extra
 from cognee.tasks.chunk_naive_llm_classifier.chunk_naive_llm_classifier import chunk_naive_llm_classifier_task
 from cognee.tasks.chunk_remove_disconnected.chunk_remove_disconnected import chunk_remove_disconnected_task
 from cognee.tasks.chunk_to_graph_decomposition.chunk_to_graph_decomposition import chunk_to_graph_decomposition_task
+from cognee.tasks.document_to_ontology.document_to_ontology import document_to_ontology
 from cognee.tasks.save_chunks_to_store.save_chunks_to_store import save_chunks_to_store_task
 from cognee.tasks.chunk_update_check.chunk_update_check import chunk_update_check_task
 from cognee.tasks.chunks_into_graph.chunks_into_graph import \
@@ -96,19 +89,20 @@ async def cognify(datasets: Union[str, list[str]] = None, user: User = None):
             cognee_config = get_cognify_config()
             graph_config = get_graph_config()
             root_node_id = None
-
-            if graph_config.infer_graph_topology and graph_config.graph_topology_task:
-                from cognee.modules.topology.topology import TopologyEngine
-                topology_engine = TopologyEngine(infer=graph_config.infer_graph_topology)
-                root_node_id = await topology_engine.add_graph_topology(files = data)
-            elif graph_config.infer_graph_topology and not graph_config.infer_graph_topology:
-                from cognee.modules.topology.topology import TopologyEngine
-                topology_engine = TopologyEngine(infer=graph_config.infer_graph_topology)
-                await topology_engine.add_graph_topology(graph_config.topology_file_path)
-            elif not graph_config.graph_topology_task:
-                root_node_id = "ROOT"
+            #
+            # if graph_config.infer_graph_topology and graph_config.graph_topology_task:
+            #     from cognee.modules.topology.topology import TopologyEngine
+            #     topology_engine = TopologyEngine(infer=graph_config.infer_graph_topology)
+            #     root_node_id = await topology_engine.add_graph_topology(files = data)
+            # elif graph_config.infer_graph_topology and not graph_config.infer_graph_topology:
+            #     from cognee.modules.topology.topology import TopologyEngine
+            #     topology_engine = TopologyEngine(infer=graph_config.infer_graph_topology)
+            #     await topology_engine.add_graph_topology(graph_config.topology_file_path)
+            # elif not graph_config.graph_topology_task:
+            #     root_node_id = "ROOT"
 
             tasks = [
+                Task(document_to_ontology, root_node_id = root_node_id),
                 Task(source_documents_to_chunks, parent_node_id = root_node_id), # Classify documents and save them as a nodes in graph db, extract text chunks based on the document type
                 Task(chunk_to_graph_decomposition_task, topology_model = KnowledgeGraph, task_config = { "batch_size": 10 }), # Set the graph topology for the document chunk data
                 Task(chunks_into_graph_task, graph_model = KnowledgeGraph, collection_name = "entities"), # Generate knowledge graphs from the document chunks and attach it to chunk nodes
