--- cognee/modules/cognify/vector/__init__.py
@@ -1,1 +1,0 @@
-from .save_data_chunks import save_data_chunks
