--- cognee/modules/data/extraction/knowledge_graph/extract_knowledge_graph.py
@@ -1,4 +1,0 @@
-from .extract_content_graph import extract_content_graph
-
-async def extract_knowledge_graph(text: str, cognitive_layer, graph_model):
-    return await extract_content_graph(text, cognitive_layer, graph_model)
