--- cognee/modules/topology/example_data.json
@@ -1,25 +1,0 @@
-[
-    {
-        "node_id": "062c22dfd99b599f90cd2d325c8bcf69",
-        "name": "062c22df-d99b-599f-90cd-2d325c8bcf69",
-        "default_relationship": {
-            "type": "related_to",
-            "source": "062c22dfd99b599f90cd2d325c8bcf69",
-            "target": "6dfe01b607d25b7783c81d6c11ce2aa7"
-        },
-        "children": [
-            {
-                "node_id": "6dfe01b607d25b7783c81d6c11ce2aa7",
-                "name": "6dfe01b6-07d2-5b77-83c8-1d6c11ce2aa7",
-                "default_relationship": {
-                    "type": "related_to",
-                    "source": "6dfe01b6-07d2-5b77-83c8-1d6c11ce2aa7",
-                    "target": "a27bb4fa897e53a594cab446e1d33dbf"
-                },
-                "children": []
-            }
-        ]
-    }
-]
-
-
