--- cognee/modules/topology/explanations/bab90046-1d9b-598c-8711-dab30f501915.txt
@@ -1,1 +1,0 @@
-German novels are fun to read and talk about nature
