+++ docs/index.md
@@ -1,58 +1,381 @@
 # cognee 
 
 
+
 ####  Deterministic LLMs Outputs for AI Engineers
+
 
 
 _Open-source framework for loading and structuring LLM context to create accurate and explainable AI solutions using knowledge graphs and vector stores_
 
 
+
 ---
+
 
 
 [![Twitter Follow](https://img.shields.io/twitter/follow/tricalt?style=social)](https://twitter.com/tricalt)
 
+
 [![Downloads](https://img.shields.io/pypi/dm/cognee.svg)](https://pypi.python.org/pypi/cognee)
+
 
 [![Star on GitHub](https://img.shields.io/github/stars/topoteretes/cognee.svg?style=social)](https://github.com/topoteretes/cognee)
 
+
 ### Let's learn about cogneeHub!
+
 
 cogneeHub is a free, open-source learning platform for those interested in creating deterministic LLM outputs. We help developers by using graphs, LLMs, and adding vector retrieval to their Machine Learning stack.
 
+
 - **Get started** — [Get started with cognee quickly and try it out for yourself.](quickstart.md)
+
 - **Conceptual Overview** — Learn about the [core concepts](conceptual_overview.md) of cognee and how it fits into your projects.
+
 - **Data Engineering and LLMOps** — Learn about some [data engineering and llmops](data_engineering_llm_ops.md) core concepts that will help you build better AI apps.
+
 - **RAGs** — We provide easy-to-follow [learning materials](rags.md) to help you learn about RAGs.
+
 - **Research** — A list of resources to help you learn more about [cognee and LLM memory research](research.md)
+
 - **Blog** — A blog where you can read about the [latest news and updates](blog/index.md) about cognee.
+
 - **Support** — [Book time](https://www.cognee.ai/#bookTime) with our team.
 
-[//]: # (- **Case Studies** — Read about [case studies]&#40;case_studies.md&#41; that show how cognee can be used in real-world applications.)
+
+[//]: # (- **Case Studies** — Read about [case studies](case_studies.md) that show how cognee can be used in real-world applications.)
 
 
 ### Vision
 
+
 ![Vision](img/roadmap.png)
+
 
 
 ### Architecture
 
+
 ![Architecture](img/architecture.png)
+
 ### Why use cognee?
+
 
 The question of using cognee is fundamentally a question of why to have deterministic outputs for your llm workflows.
 
+
 1. **Cost-effective** — cognee extends the capabilities of your LLMs without the need for expensive data processing tools.
+
 
 2. **Self-contained** — cognee runs as a library and is simple to use
 
+
 3. **Interpretable** — Navigate graphs instead of embeddings to understand your data.
+
 
 4. **User Guided** —  cognee lets you control your input and provide your own Pydantic data models 
 
 
 
+
 ## License
 
+
 This project is licensed under the terms of the Apache License 2.0.
+
+
+[//]: # (<style>)
+
+[//]: # ()
+[//]: # (.container {)
+
+[//]: # ()
+[//]: # (  display: flex;)
+
+[//]: # ()
+[//]: # (  justify-content: space-around;)
+
+[//]: # ()
+[//]: # (  margin-top: 20px;)
+
+[//]: # ()
+[//]: # (})
+
+[//]: # ()
+[//]: # ()
+[//]: # (.container div {)
+
+[//]: # ()
+[//]: # (  width: 28%;)
+
+[//]: # ()
+[//]: # (  padding: 20px;)
+
+[//]: # ()
+[//]: # (  box-sizing: border-box;)
+
+[//]: # ()
+[//]: # (  border: 1px solid #e0e0e0;)
+
+[//]: # ()
+[//]: # (  border-radius: 8px;)
+
+[//]: # ()
+[//]: # (  background-color: #f9f9f9;)
+
+[//]: # ()
+[//]: # (})
+
+[//]: # ()
+[//]: # ()
+[//]: # (.container h2 {)
+
+[//]: # ()
+[//]: # (  font-size: 1.25em;)
+
+[//]: # ()
+[//]: # (  margin-bottom: 10px;)
+
+[//]: # ()
+[//]: # (})
+
+[//]: # ()
+[//]: # ()
+[//]: # (.container p {)
+
+[//]: # ()
+[//]: # (  margin-bottom: 20px;)
+
+[//]: # ()
+[//]: # (  line-height: 1.6;)
+
+[//]: # ()
+[//]: # (})
+
+[//]: # ()
+[//]: # ()
+[//]: # (.button-container {)
+
+[//]: # ()
+[//]: # (  text-align: center;)
+
+[//]: # ()
+[//]: # (  margin: 30px 0;)
+
+[//]: # ()
+[//]: # (})
+
+[//]: # ()
+[//]: # ()
+[//]: # (.button-container a {)
+
+[//]: # ()
+[//]: # (  display: inline-block;)
+
+[//]: # ()
+[//]: # (  padding: 15px 25px;)
+
+[//]: # ()
+[//]: # (  background-color: #007bff;)
+
+[//]: # ()
+[//]: # (  color: white;)
+
+[//]: # ()
+[//]: # (  text-decoration: none;)
+
+[//]: # ()
+[//]: # (  border-radius: 5px;)
+
+[//]: # ()
+[//]: # (  font-size: 1em;)
+
+[//]: # ()
+[//]: # (})
+
+[//]: # ()
+[//]: # ()
+[//]: # (.button-container a:hover {)
+
+[//]: # ()
+[//]: # (  background-color: #0056b3;)
+
+[//]: # ()
+[//]: # (})
+
+[//]: # ()
+[//]: # ()
+[//]: # (.resources {)
+
+[//]: # ()
+[//]: # (  margin-top: 40px;)
+
+[//]: # ()
+[//]: # (})
+
+[//]: # ()
+[//]: # ()
+[//]: # (.resources h2 {)
+
+[//]: # ()
+[//]: # (  font-size: 1.5em;)
+
+[//]: # ()
+[//]: # (  margin-bottom: 20px;)
+
+[//]: # ()
+[//]: # (})
+
+[//]: # ()
+[//]: # ()
+[//]: # (.resources ul {)
+
+[//]: # ()
+[//]: # (  list-style: none;)
+
+[//]: # ()
+[//]: # (  padding: 0;)
+
+[//]: # ()
+[//]: # (})
+
+[//]: # ()
+[//]: # ()
+[//]: # (.resources li {)
+
+[//]: # ()
+[//]: # (  margin-bottom: 10px;)
+
+[//]: # ()
+[//]: # (})
+
+[//]: # ()
+[//]: # ()
+[//]: # (.resources a {)
+
+[//]: # ()
+[//]: # (  color: #007bff;)
+
+[//]: # ()
+[//]: # (  text-decoration: none;)
+
+[//]: # ()
+[//]: # (})
+
+[//]: # ()
+[//]: # ()
+[//]: # (.resources a:hover {)
+
+[//]: # ()
+[//]: # (  text-decoration: underline;)
+
+[//]: # ()
+[//]: # (})
+
+[//]: # ()
+[//]: # (</style>)
+
+[//]: # ()
+[//]: # ()
+[//]: # (# New to cognee?)
+
+[//]: # ()
+[//]: # ()
+[//]: # (The getting started guide covers adding a GraphRAG data store to your AI app, sending events, identifying users, extracting actions and insights, and interconnecting separate datasets.)
+
+[//]: # ()
+[//]: # ()
+[//]: # (<div class="button-container">)
+
+[//]: # ()
+[//]: # (  <a href="./quickstart.md">Get started</a>)
+
+[//]: # ()
+[//]: # (</div>)
+
+[//]: # ()
+[//]: # ()
+[//]: # (<div class="container">)
+
+[//]: # ()
+[//]: # (  <div>)
+
+[//]: # ()
+[//]: # (    <h2>Ingest Data</h2>)
+
+[//]: # ()
+[//]: # (    <p>Learn how to manage ingestion of events, customer data or third party data for use with cognee.</p>)
+
+[//]: # ()
+[//]: # (    <a href="#">Explore</a>)
+
+[//]: # ()
+[//]: # (  </div>)
+
+[//]: # ()
+[//]: # (  <div>)
+
+[//]: # ()
+[//]: # (    <h2>Templates</h2>)
+
+[//]: # ()
+[//]: # (    <p>Analyze and enrich your data and improve LLM answers with a series of templates using cognee tasks and pipelines.</p>)
+
+[//]: # ()
+[//]: # (    <a href="#">Browse templates</a>)
+
+[//]: # ()
+[//]: # (  </div>)
+
+[//]: # ()
+[//]: # (  <div>)
+
+[//]: # ()
+[//]: # (    <h2>API</h2>)
+
+[//]: # ()
+[//]: # (    <p>Push or pull data to build custom functionality or create bespoke views for your business needs.</p>)
+
+[//]: # ()
+[//]: # (    <a href="#">Explore</a>)
+
+[//]: # ()
+[//]: # (  </div>)
+
+[//]: # ()
+[//]: # (</div>)
+
+[//]: # ()
+[//]: # ()
+[//]: # (<div class="resources">)
+
+[//]: # ()
+[//]: # (  <h2>Resources</h2>)
+
+[//]: # ()
+[//]: # (  <ul>)
+
+[//]: # ()
+[//]: # (    <li><a href="#">What is GraphRAG</a></li>)
+
+[//]: # ()
+[//]: # (    <li><a href="#">Research</a></li>)
+
+[//]: # ()
+[//]: # (    <li><a href="#">Community</a></li>)
+
+[//]: # ()
+[//]: # (    <li><a href="#">Community</a></li>)
+
+[//]: # ()
+[//]: # (    <li><a href="#">API Reference</a></li>)
+
+[//]: # ()
+[//]: # (    <li><a href="#">Support</a></li>)
+
+[//]: # ()
+[//]: # (  </ul>)
+
+[//]: # ()
+[//]: # (</div>)
